+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.20.0
 
-- `rateLimit` accepts multiple `RateLimiter` objects. Interfaces have changed slightly but existing code will work. The `getStats` method now returns an array of stats objects.
+- `rateLimit` accepts multiple `RateLimiter` objects.
+   - A new `RateLimiter` interface has been created in place of `RateLimitOptions`.
+   - The `getStats` method now returns an array of stats objects.
 
 # 0.19.0