+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.20.0
 
 - `rateLimit` accepts multiple `RateLimiter` objects.
-   - A new `RateLimiter` interface has been created in place of `RateLimitOptions`.
+   - A new `RateLimiter` interface has been created in place of `RateLimitOptions`. The `maxItemsPerPeriod` field is now required.
    - The `getStats` method now returns an array of stats objects.
 
 # 0.19.0